requirements.txt

matplotlib==3.8.3
opencv_python==4.9.0.80
plot_keras_history==1.1.38
tensorflow==2.16.1
fastapi==0.111.0
keras==3.2.1
numpy==1.26.4
Pillow==10.3.0
prometheus_client==0.20.0
uvicorn==0.29.0
python-multipart
tensorflow
